+++ docs/changelog.md
@@ -1,2 +1,15 @@
 # Changelog
 
+## Version 0.0.2
+
+### Bug fixes:
+
+Fixing a bug in time series LocalZScoreGenerator
+
+### New features
+
+GaussianNoiseClassesGenerator: adds Gaussian white noise to classes separately
+
+## Version 0.0.1
+
+This is the very first release of badgers.